requirements.txt
avro-python3==1.10.2
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
pyOpenSSL==23.0.0
python-snappy==0.6.1
Werkzeug==2.2.3
